requirements.txt
pandas
python-docx
openai
spacy
scikit-learn
chromadb
streamlit
python-dotenv
pdfplumber
